# engine/src/main/java/pl/poznan/put/rnatangoengine/dto/Chain.java
package pl.poznan.put.rnatangoengine.dto;

import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
import com.fasterxml.jackson.databind.annotation.JsonSerialize;
import org.immutables.value.Value;

@Value.Immutable
@Value.Style(privateNoargConstructor = true)
@JsonSerialize(as = ImmutableChain.class)
@JsonDeserialize(as = ImmutableChain.class)
public interface Chain {
  String name();

  String sequence();
}