# openapi.yml
openapi: 3.0.3
info:
  title: RNAtango
  description: This is the RNAtango server based on the OpenAPI 3.0 specification.
  contact:
    email: tomasz.zok@cs.put.poznan.pl
  license:
    name: MIT
    url: https://opensource.org/license/mit/
  version: '1'
servers:
  - url: https://rnatango.cs.put.poznan.pl/api
paths:
...
  /single:
    post:
      summary: Analyze a single file
      description: Perform analysis of a single file specifying a set of chain names and nucleotide ranges. The result contains tabular data with torsion angles.
      operationId: postSingle
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SingleInput'
        required: true
      responses:
        '202':
          description: Your request has been accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskIdResponse'
  /single/{taskId}:
    get:
      summary: Check status of single file analysis
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            $ref: '#/components/schemas/TaskId'
      responses:
        '200':
          description: The current status of the single file analysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '404':
          description: Request does not exist
  /single/{taskId}/result:
    get:
      summary: Get the result of single file analysis
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            $ref: '#/components/schemas/TaskId'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleOutput'
        '404':
          description: Request does not exist
...
components:
  schemas:
    TaskId:
      type: string
      example: 9ca7c1c2-31e8-11ee-be56-0242ac120002
    TaskIdResponse:
      type: object
      properties:
        taskId:
          $ref: '#/components/schemas/TaskId'
      required:
        - taskId
    Status:
      type: string
      enum:
        - WAITING
        - PROCESSING
        - SUCCESS
        - FAILED
    StatusResponse:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/Status'
        resultUrl:
          type: string
      required:
        - status
        - resultUrl
    FileId:
      type: string
      example: 9ca7c1c2-31e8-11ee-be56-0242ac120002
    NucleotideRange:
      type: object
      properties:
        fromInclusive:
          type: integer
          example: 0
        toInclusive:
          type: integer
          example: 10
      required:
        - fromInclusive
        - toInclusive
    SelectedChain:
      type: object
      properties:
        name:
          type: string
          example: "A"
        nucleotideRange:
          $ref: '#/components/schemas/NucleotideRange'
    Selection:
      type: object
      properties:
        modelName:
          type: string
          example: 1
        chains:
          type: array
          items:
            $ref: '#/components/schemas/SelectedChain'
      required:
        - chain
        - nucleotideRange
    SingleInput:
      type: object
      properties:
        fileId:
          $ref: '#/components/schemas/FileId'
        selections:
          type: array
          items:
            $ref: '#/components/schemas/Selection'
          minItems: 1
      required:
        - fileId
        - selections
...
    Chain:
      type: object
      properties:
        name:
          type: string
          example: 1
        sequence:
          type: string
          example: AAAATTATTACCATT
      required:
        - name
        - sequence
...
    Residue:
      type: object
      properties:
        name:
          type: string
          example: G
        number:
          type: integer
          example: 1
        icode:
          type: string
          default: null
          example: null
        torsionAngles:
          type: array
          items:
            $ref: '#/components/schemas/TorsionAngleInResidue'
      required:
        - name
        - number
    TorsionAngleInResidue:
      description: An array of torsion angle values in degrees for one residue given in order alpha, beta, gamma, delta, epsilon, zeta, chi, eta, theta, eta', theta'.
      type: array
      items:
        type: object
        properties:
          angle:
            $ref: '#/components/schemas/Angle'
          value:
            type: number
            format: double
      minItems: 11
      maxItems: 11
      example:
        - {angle: ALPHA,value: null}
        - {angle: BETA,value: 25}
        - {angle: GAMMA,value: -14}
        - {angle: DELTA,value: 45}
        - {angle: EPSILON,value: 100}
        - {angle: ZETA,value: 41}
        - {angle: CHI,value: 47}
        - {angle: ETA,value: 57}
        - {angle: THETA,value: 57}
        - {angle: ETA_PRIM,value: 2}
        - {angle: THETA_PRIM,value: 44}
    TorsionAnglesInChain:
      type: object
      properties:
        chain:
          $ref: '#/components/schemas/Chain'
        residues:
          type: array
          items:
            $ref: '#/components/schemas/Residue'
      required:
        - chain
        - residues
    SingleOutput:
      type: object
      properties:
        torsionAngles:
          type: array
          items:
            $ref: '#/components/schemas/TorsionAnglesInChain'
      required:
        - torsionAngles
    Angle:
      type: string
      enum:
        - ALPHA
        - BETA
        - GAMMA
        - DELTA
        - EPSILON
        - ZETA
        - CHI
        - ETA
        - THETA
        - ETA_PRIM
        - THETA_PRIM
...